升级引擎

+++ libs/core/extension/resource/Resource.js
@@ -1,20 +1,31 @@
-/**
- * Copyright (c) Egret-Labs.org. Permission is hereby granted, free of charge,
- * to any person obtaining a copy of this software and associated documentation
- * files (the "Software"), to deal in the Software without restriction, including
- * without limitation the rights to use, copy, modify, merge, publish, distribute,
- * sublicense, and/or sell copies of the Software, and to permit persons to whom
- * the Software is furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included
- * in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
- * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR
- * PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE
- * FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE,
- * ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
+//////////////////////////////////////////////////////////////////////////////////////
+//
+//  Copyright (c) 2014-2015, Egret Technology Inc.
+//  All rights reserved.
+//  Redistribution and use in source and binary forms, with or without
+//  modification, are permitted provided that the following conditions are met:
+//
+//     * Redistributions of source code must retain the above copyright
+//       notice, this list of conditions and the following disclaimer.
+//     * Redistributions in binary form must reproduce the above copyright
+//       notice, this list of conditions and the following disclaimer in the
+//       documentation and/or other materials provided with the distribution.
+//     * Neither the name of the Egret nor the
+//       names of its contributors may be used to endorse or promote products
+//       derived from this software without specific prior written permission.
+//
+//  THIS SOFTWARE IS PROVIDED BY EGRET AND CONTRIBUTORS "AS IS" AND ANY EXPRESS
+//  OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
+//  OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED.
+//  IN NO EVENT SHALL EGRET AND CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+//  INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+//  LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;LOSS OF USE, DATA,
+//  OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+//  LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+//  NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
+//  EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+//
+//////////////////////////////////////////////////////////////////////////////////////
 var RES;
 (function (RES) {
     /**
@@ -23,6 +34,7 @@ var RES;
      * @param url {string} 配置文件路径(resource.json的路径)
      * @param resourceRoot {string} 资源根路径。配置中的所有url都是这个路径的相对值。最终url是这个字符串与配置里资源项的url相加的值。
      * @param type {string} 配置文件的格式。确定要用什么解析器来解析配置文件。默认"json"
+     * @see #setMaxRetryTimes
      */
     function loadConfig(url, resourceRoot, type) {
         if (resourceRoot === void 0) { resourceRoot = ""; }
@@ -36,6 +48,7 @@ var RES;
      * @param name {string} 要加载资源组的组名
      * @param priority {number} 加载优先级,可以为负数,默认值为0。
      * 低优先级的组必须等待高优先级组完全加载结束才能开始，同一优先级的组会同时加载。
+     * @see #setMaxRetryTimes
      */
     function loadGroup(name, priority) {
         if (priority === void 0) { priority = 0; }
@@ -47,6 +60,7 @@ var RES;
      * @method RES.isGroupLoaded
      * @param name {string} 组名
      * @returns {boolean}
+     * @see #setMaxRetryTimes
      */
     function isGroupLoaded(name) {
         return instance.isGroupLoaded(name);
@@ -57,6 +71,7 @@ var RES;
      * @method RES.getGroupByName
      * @param name {string} 组名
      * @returns {egret.ResourceItem}
+     * @see #setMaxRetryTimes
      */
     function getGroupByName(name) {
         return instance.getGroupByName(name);
@@ -70,6 +85,7 @@ var RES;
      * @param keys {egret.Array<string>} 要包含的键名列表，key对应配置文件里的name属性或sbuKeys属性的一项或一个资源组名。
      * @param override {boolean} 是否覆盖已经存在的同名资源组,默认false。
      * @returns {boolean}
+     * @see #setMaxRetryTimes
      */
     function createGroup(name, keys, override) {
         if (override === void 0) { override = false; }
@@ -81,6 +97,7 @@ var RES;
      * @method RES.hasRes
      * @param key {string} 对应配置文件里的name属性或sbuKeys属性的一项。
      * @returns {boolean}
+     * @see #setMaxRetryTimes
      */
     function hasRes(key) {
         return instance.hasRes(key);
@@ -91,6 +108,7 @@ var RES;
      * @method RES.parseConfig
      * @param data {any} 配置文件数据，请参考resource.json的配置文件格式。传入对应的json对象即可。
      * @param folder {string} 加载项的路径前缀。
+     * @see #setMaxRetryTimes
      */
     function parseConfig(data, folder) {
         if (folder === void 0) { folder = ""; }
@@ -102,6 +120,7 @@ var RES;
      * @method RES.getRes
      * @param key {string} 对应配置文件里的name属性或sbuKeys属性的一项。
      * @returns {any}
+     * @see #setMaxRetryTimes
      */
     function getRes(key) {
         return instance.getRes(key);
@@ -113,6 +132,7 @@ var RES;
      * @param key {string} 对应配置文件里的name属性或sbuKeys属性的一项。
      * @param compFunc {Function} 回调函数。示例：compFunc(data,key):void。
      * @param thisObject {any} 回调函数的this引用
+     * @see #setMaxRetryTimes
      */
     function getResAsync(key, compFunc, thisObject) {
         instance.getResAsync(key, compFunc, thisObject);
@@ -125,6 +145,7 @@ var RES;
      * @param compFunc {Function} 回调函数。示例：compFunc(data,url):void。
      * @param thisObject {any} 回调函数的this引用
      * @param type {string} 文件类型(可选)。请使用ResourceItem类中定义的静态常量。若不设置将根据文件扩展名生成。
+     * @includeExample extension/resource/GetResByUrl.ts
      */
     function getResByUrl(url, compFunc, thisObject, type) {
         if (type === void 0) { type = ""; }
@@ -135,16 +156,19 @@ var RES;
      * 销毁单个资源文件或一组资源的缓存数据,返回是否删除成功。
      * @method RES.destroyRes
      * @param name {string} 配置文件中加载项的name属性或资源组名
+     * @param force {boolean} 销毁一个资源组时其他资源组有同样资源情况资源是否会被删除，默认值true
      * @returns {boolean}
+     * @see #setMaxRetryTimes
      */
-    function destroyRes(name) {
-        return instance.destroyRes(name);
+    function destroyRes(name, force) {
+        return instance.destroyRes(name, force);
     }
     RES.destroyRes = destroyRes;
     /**
      * 设置最大并发加载线程数量，默认值是2.
      * @method RES.setMaxLoadingThread
      * @param thread {number} 要设置的并发加载数。
+     * @see #setMaxRetryTimes
      */
     function setMaxLoadingThread(thread) {
         instance.setMaxLoadingThread(thread);
@@ -153,6 +177,7 @@ var RES;
     /**
      * 设置资源加载失败时的重试次数，默认值是 3。
      * @param retry 要设置的重试次数。
+     * @includeExample extension/resource/Resource.ts
      */
     function setMaxRetryTimes(retry) {
         instance.setMaxRetryTimes(retry);
@@ -170,6 +195,7 @@ var RES;
      * 要在所有三个阶段都侦听事件，请调用 addEventListener 两次：一次将 useCapture 设置为 true，一次将 useCapture 设置为 false。
      * @param priority {number} 事件侦听器的优先级。优先级由一个带符号的 32 位整数指定。数字越大，优先级越高。优先级为 n 的所有侦听器会在
      * 优先级为 n -1 的侦听器之前得到处理。如果两个或更多个侦听器共享相同的优先级，则按照它们的添加顺序进行处理。默认优先级为 0。
+     * @see #setMaxRetryTimes
      */
     function addEventListener(type, listener, thisObject, useCapture, priority) {
         if (useCapture === void 0) { useCapture = false; }
@@ -184,12 +210,16 @@ var RES;
      * @param listener {Function} 侦听函数
      * @param thisObject {any} 侦听函数绑定的this对象
      * @param useCapture {boolean} 是否使用捕获，这个属性只在显示列表中生效。
+     * @see #setMaxRetryTimes
      */
     function removeEventListener(type, listener, thisObject, useCapture) {
         if (useCapture === void 0) { useCapture = false; }
         instance.removeEventListener(type, listener, thisObject, useCapture);
     }
     RES.removeEventListener = removeEventListener;
+    /**
+     * @private
+     */
     var Resource = (function (_super) {
         __extends(Resource, _super);
         /**
@@ -217,6 +247,7 @@ var RES;
              * 异步获取资源参数缓存字典
              */
             this.asyncDic = {};
+            this._loadedUrlTypes = {};
             this.init();
         }
         var __egretProto__ = Resource.prototype;
@@ -454,14 +485,14 @@ var RES;
                 name = RES.AnalyzerBase.getStringPrefix(key);
                 type = this.resConfig.getType(name);
                 if (type == "") {
-                    compFunc.call(thisObject, null);
+                    egret.__callAsync(compFunc, thisObject);
                     return;
                 }
             }
             var analyzer = this.getAnalyzerByType(type);
             var res = analyzer.getRes(key);
             if (res) {
-                compFunc.call(thisObject, res);
+                egret.__callAsync(compFunc, thisObject, res, key);
                 return;
             }
             var args = { key: key, compFunc: compFunc, thisObject: thisObject };
@@ -485,16 +516,20 @@ var RES;
         __egretProto__.getResByUrl = function (url, compFunc, thisObject, type) {
             if (type === void 0) { type = ""; }
             if (!url) {
-                compFunc.call(thisObject, null);
+                egret.__callAsync(compFunc, thisObject);
                 return;
             }
             if (!type)
                 type = this.getTypeByUrl(url);
+            if (this._loadedUrlTypes[url] != null && this._loadedUrlTypes[url] != type) {
+                egret.$warn(2002);
+            }
+            this._loadedUrlTypes[url] = type;
             var analyzer = this.getAnalyzerByType(type);
             var name = url;
             var res = analyzer.getRes(name);
             if (res) {
-                compFunc.call(thisObject, res);
+                egret.__callAsync(compFunc, thisObject, res, url);
                 return;
             }
             var args = { key: name, compFunc: compFunc, thisObject: thisObject };
@@ -570,11 +605,13 @@ var RES;
          * 销毁单个资源文件或一组资源的缓存数据,返回是否删除成功。
          * @method RES.destroyRes
          * @param name {string} 配置文件中加载项的name属性或资源组名
+         * @param force {boolean} 销毁一个资源组时其他资源组有同样资源情况资源是否会被删除，默认值true
          * @returns {boolean}
          */
-        __egretProto__.destroyRes = function (name) {
+        __egretProto__.destroyRes = function (name, force) {
+            if (force === void 0) { force = true; }
             var group = this.resConfig.getRawGroupByName(name);
-            if (group) {
+            if (group && group.length > 0) {
                 var index = this.loadedGroups.indexOf(name);
                 if (index != -1) {
                     this.loadedGroups.splice(index, 1);
@@ -582,21 +619,68 @@ var RES;
                 var length = group.length;
                 for (var i = 0; i < length; i++) {
                     var item = group[i];
-                    item.loaded = false;
-                    var analyzer = this.getAnalyzerByType(item.type);
-                    analyzer.destroyRes(item.name);
+                    if (!force && this.isResInLoadedGroup(item.name)) {
+                    }
+                    else {
+                        item.loaded = false;
+                        var analyzer = this.getAnalyzerByType(item.type);
+                        analyzer.destroyRes(item.name);
+                        this.removeLoadedGroupsByItemName(item.name);
+                    }
                 }
                 return true;
             }
             else {
                 var type = this.resConfig.getType(name);
-                if (type == "")
-                    return false;
+                if (type == "") {
+                    type = this._loadedUrlTypes[name];
+                    if (type == null || type == "") {
+                        return false;
+                    }
+                    delete this._loadedUrlTypes[name];
+                    var analyzer = this.getAnalyzerByType(type);
+                    analyzer.destroyRes(name);
+                    return true;
+                }
                 item = this.resConfig.getRawResourceItem(name);
                 item.loaded = false;
                 analyzer = this.getAnalyzerByType(type);
-                return analyzer.destroyRes(name);
+                var result = analyzer.destroyRes(name);
+                this.removeLoadedGroupsByItemName(item.name);
+                return result;
             }
+        };
+        __egretProto__.removeLoadedGroupsByItemName = function (name) {
+            var loadedGroups = this.loadedGroups;
+            var loadedGroupLength = loadedGroups.length;
+            for (var i = 0; i < loadedGroupLength; i++) {
+                var group = this.resConfig.getRawGroupByName(loadedGroups[i]);
+                var length = group.length;
+                for (var j = 0; j < length; j++) {
+                    var item = group[j];
+                    if (item.name == name) {
+                        loadedGroups.splice(i, 1);
+                        i--;
+                        loadedGroupLength = loadedGroups.length;
+                        break;
+                    }
+                }
+            }
+        };
+        __egretProto__.isResInLoadedGroup = function (name) {
+            var loadedGroups = this.loadedGroups;
+            var loadedGroupLength = loadedGroups.length;
+            for (var i = 0; i < loadedGroupLength; i++) {
+                var group = this.resConfig.getRawGroupByName(loadedGroups[i]);
+                var length = group.length;
+                for (var j = 0; j < length; j++) {
+                    var item = group[j];
+                    if (item.name == name) {
+                        return true;
+                    }
+                }
+            }
+            return false;
         };
         /**
          * 设置最大并发加载线程数量，默认值是2.
